requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
python-dotenv
cryptography
google-generativeai
Werkzeug
gunicorn
requests
